Create Image of photoevaporation of the protoplanetary disk
Protoplanetary disks are the site of planet formation, and they are composed of gas and dust. The process called photovaporation may affect the evolution of the disk. In this process, the gas is excited by the irradiation of high-energy photons from the central star and released from the stellar gravity and evaporates into the outer space.